backend/requirements.txt
Django>=4.2,<5.0
scalekit-sdk-python>=2.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
django-cors-headers>=4.3.0
